requirements.txt
ultralytics
opencv-python
Pillow
PyYAML
pytest
